refactor(metadata/sqlite): split type definitaions into indivitial files
