+++ src/database/metadata/types/group.rs
@@ -1,35 +1,6 @@
 use chrono::DateTime;
 use chrono::Utc;
 
-#[derive(sqlx::FromRow)]
-pub struct Entity {
-    pub id: i64,
-    pub name: String,
-    pub description: String,
-
-    pub create_time: DateTime<Utc>,
-    pub modify_time: DateTime<Utc>,
-}
-
-pub enum GetEntityOpt {
-    Id(i64),
-    Name(String),
-}
-
-pub enum ListEntityOpt {
-    /// return all rows from DB.
-    All,
-    /// return rows which id in the id list from DB.
-    Ids(Vec<i64>),
-}
-
-#[derive(sqlx::Type, Default, PartialEq, Debug, Clone)]
-pub enum Category {
-    #[default]
-    Batch,
-    Stream,
-}
-
 #[derive(sqlx::FromRow, Default, Clone)]
 pub struct Group {
     pub id: i64,
@@ -71,4 +42,11 @@ pub enum ListGroupOpt {
     All,
     /// return rows which id in the id list from DB.
     Ids(Vec<i64>),
+}
+
+#[derive(sqlx::Type, Default, PartialEq, Debug, Clone)]
+pub enum Category {
+    #[default]
+    Batch,
+    Stream,
 }
