+++ src/database/metadata/types/mod.rs
@@ -1,19 +1,19 @@
+pub mod entity;
 pub mod feature;
-pub mod feature_value_type;
-pub mod types;
+pub mod group;
 
-pub use types::Category;
-pub use types::CreateGroupOpt;
-pub use types::GetGroupOpt;
-pub use types::Group;
-pub use types::ListGroupOpt;
+pub use group::Category;
+pub use group::CreateGroupOpt;
+pub use group::GetGroupOpt;
+pub use group::Group;
+pub use group::ListGroupOpt;
 
-pub use types::Entity;
-pub use types::GetEntityOpt;
-pub use types::ListEntityOpt;
+pub use entity::Entity;
+pub use entity::GetEntityOpt;
+pub use entity::ListEntityOpt;
 
 pub use feature::CreateFeatureOpt;
 pub use feature::Feature;
 pub use feature::GetFeatureOpt;
 pub use feature::ListFeatureOpt;
-pub use feature_value_type::FeatureValueType;
+pub use feature::FeatureValueType;
